Update x86 dependency to 0.45.0
Older versions of the x86 crate used the `llvm_asm!` macro, which was removed in the latest Rust nightly builds, from 2022-01-17 onwards. The latest x86 crate version, 0.45.0, fixes this problem. As this crate depends on older versions of the x86 crate, it is broken on the affected Rust nightly builds.

To fix this situation, update the x86 dependency to its latest version.