Update phf to 0.9.0
The x86 dependency already pulls the newer version of phf, and as far as
I can see this crate can also use that version without any source code
changes. Having both crates depend on the same version of phf is good
because it reduces build times (it is no longer necessary to download
and compile two versions of the library) and potentially executable
sizes.